Moved header files into a directory that matches their namespaces.
