+++ ros/drake_ros_systems/include/drake/ros/ros_sensor_publisher_odometry.h
