+++ ros/drake_ros_systems/include/drake/ros/ros_tf_publisher.h
