+++ ros/drake_ros_systems/include/drake/ros/ros_vehicle_system.h
